More refactoring for code reuse
